+++ packages/patrol/darwin/patrol/Sources/PatrolImpl/AutomatorServer/Contracts.swift
@@ -488,6 +488,10 @@ public struct AndroidTakeCameraPhotoRequest: Codable {
   public var timeoutMillis: Int?
 }
 
+public struct AndroidTakeNativeScreenshotRequest: Codable {
+  public var tag: String
+}
+
 public struct IOSTakeCameraPhotoRequest: Codable {
   public var shutterButtonSelector: IOSSelector?
   public var doneButtonSelector: IOSSelector?